% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conductor_cars.R
\docType{data}
\name{conductor_cars}
\alias{conductor_cars}
\title{Conductor conductor_cars}
\format{
Una data frame amb 5 files i 3 variables
\describe{
\item{camp}{camp (Sepal.Width Petal.Length  Petal.Width Sepal.Length Species )}
\item{descripcio}{descripcio (AmplitudPetal   AmplitudSepal)}
\item{exclusio}{exclusio ()}
}
}
\usage{
conductor_cars
}
\description{
Conductor conductor_cars
}
\keyword{datasets}
