% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_flow_chart_diagrama.R
\name{criteris_exclusio_diagrama}
\alias{criteris_exclusio_diagrama}
\title{Dibuixa un Flow Chart Diagrama a partir del Conductor}
\usage{
criteris_exclusio_diagrama(
  dt = dades,
  taulavariables = "VARIABLES_R3b.xls",
  criteris = "exclusio",
  pob_lab = c("Pob inicial", "Pob final"),
  etiquetes = "etiqueta_exclusio",
  ordre = "exc_ordre",
  grups = NA,
  sequencial = F,
  colors = c("white", "grey"),
  forma = c("ellipse", "box"),
  missings = T,
  ...
)
}
\arguments{
\item{dt}{Dt}

\item{taulavariables}{Taulavariables}

\item{criteris}{Criteris}

\item{pob_lab}{Pob labels}

\item{etiquetes}{Etiquetes}

\item{ordre}{Ordre}

\item{grups}{Grups}

\item{sequencial}{Sequencial}

\item{colors}{Colors}

\item{forma}{Forma}

\item{missings}{missings}

\item{...}{Altres parametres}
}
\description{
Dibuixa un Flow Chart Diagrama a partir del Conductor
}
\examples{

conductor_cars
exemple

k2<-criteris_exclusio_diagrama(
dt=exemple,
taulavariables=conductor_cars,
criteris="exclusio",
ordre="exc_ordre",
grups="lloc",
etiquetes="descripcio",
sequencial = TRUE,
pob_lab=c("Pob inicial","Pob final"),
colors=c("white","grey"),
forma=c("ellipse","box"))

k2

}
