% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generar_flow_chart_consort.R
\name{criteris_exclusio_ggconsort}
\alias{criteris_exclusio_ggconsort}
\title{Aplica criteris partir d una llista de n criteris exclusio}
\usage{
criteris_exclusio_ggconsort(
  dt = "dades",
  taulavariables = "VARIABLES_R3b.xls",
  criteris = "exclusio1",
  missings = T,
  ...
)
}
\arguments{
\item{dt}{Dataframe tibble}

\item{taulavariables}{String referent a path o tibble on hi ha una columna consten les exclusions: "conductor_cars.xls"}

\item{criteris}{String referent a la columna on consten les exclusions}

\item{missings}{Logic TRUE FALSE  (De moment no operatiu)}

\item{...}{Altres prametres}
}
\description{
Aplica criteris partir d una llista de n criteris exclusio que estan en un fitxer extern (Tipo excel)
}
