% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_flow_chart_diagrama.R
\name{diagramaFlowchart}
\alias{diagramaFlowchart}
\title{Dibuixa un Flow Chart Diagrama}
\usage{
diagramaFlowchart(
  grups = 3,
  pob = c(1700),
  pob_lab = c("Poblaci? Alt Pened?s"),
  pob1 = c(1000, 500),
  pob2 = c(400, 200),
  pob3 = c(300, 100),
  pob_lab1 = c("A INICIAL", "A FINAL"),
  pob_lab2 = c("B Inicial", "B Final"),
  pob_lab3 = c("C Inicial", "C Final"),
  exc1 = c(50, 300, 150),
  exc2 = c(100, 50, 50),
  exc3 = c(100, 50, 50),
  exc_lab1 = c("Edat>90 anys", "M.Cardio", "M.Pulmonar"),
  exc_lab2 = c("Edat>90 anys", "M.Cardio", "M.Pulmonar"),
  exc_lab3 = c("Edat>90 anys", "M.Cardio", "M.Pulmonar"),
  colors = c("white", "grey"),
  forma = c("ellipse", "box")
)
}
\arguments{
\item{grups}{Grups}

\item{pob}{Pob}

\item{pob_lab}{Pob_lab}

\item{pob1}{Pob1}

\item{pob2}{Pob2}

\item{pob3}{Pob3}

\item{pob_lab1}{Pob lab1}

\item{pob_lab2}{Pob lab2}

\item{pob_lab3}{Pob lab3}

\item{exc1}{Exc1}

\item{exc2}{Exc2}

\item{exc3}{Exc3}

\item{exc_lab1}{Exc lab1}

\item{exc_lab2}{Exc lab2}

\item{exc_lab3}{Exc lab3}

\item{colors}{Colors}

\item{forma}{Forma}
}
\description{
Dibuixa un Flow Chart Diagrama
}
\examples{

k<-diagramaFlowchart(
grups=1,
pob=c(1700),
pob_lab=c("Poblacio Alt Penedes"),
pob1=c(1000,500),
pob2=c(400,200),
pob3=c(300,100),
pob_lab1=c("A INICIAL","A FINAL"),
pob_lab2=c("B Inicial","B Final"),
pob_lab3=c("C Inicial","C Final"),
exc1=c(50,300,150),
exc2=c(100,50,50),
exc3=c(100,50,50),
exc_lab1=c('Edat>90 anys','M.Cardio','M.Pulmonar'),
exc_lab2=c('Edat>90 anys','M.Cardio','M.Pulmonar'),
exc_lab3=c('Edat>90 anys','M.Cardio','M.Pulmonar'),
colors=c('white','grey'),
forma=c('ellipse','box'))
k
}
