% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_flow_chart_diagrama.R
\name{diagramaFlowchart1G}
\alias{diagramaFlowchart1G}
\title{Diagrama Flowchart 1 Grup}
\usage{
diagramaFlowchart1G(
  pob_lab1 = c("Pob Inicial", "Pob Final"),
  pob1 = c(1000, 50),
  exc1 = c(10, 1),
  exc_lab1 = c("Edat>90 anys", "kkk"),
  colors = c("white", "grey"),
  forma = c("box", "box")
)
}
\arguments{
\item{pob_lab1}{pob lab1}

\item{pob1}{Pob1}

\item{exc1}{Exc1}

\item{exc_lab1}{Exc lab1}

\item{colors}{Colors}

\item{forma}{Forma}
}
\description{
Diagrama Flowchart 1 Grup
}
