% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_flow_chart_diagrama.R
\name{diagramaFlowchart3G}
\alias{diagramaFlowchart3G}
\title{Diagrama Flowchart 3 Grups}
\usage{
diagramaFlowchart3G(
  pob_lab = c("Poblaci? Total"),
  pob_lab1 = c("Poblaci? Inicial A", "Poblaci? Final A"),
  pob_lab2 = c("Poblaci? Inicial B", "Poblaci? Final B"),
  pob_lab3 = c("Poblaci? Inicial C", "Poblaci? Final C"),
  pob = c(70211123),
  pob1 = c(10088, 50),
  exc1 = c(1021, 111, 9),
  exc_lab1 = c("Edat>90 anys", "Cardio", "J"),
  pob2 = c(19002, 599),
  exc2 = c(1002, 150, 90),
  exc_lab2 = c("Edat>76 anys", "Rata", "U"),
  pob3 = c(19002, 599),
  exc3 = c(1002, 150, 0),
  exc_lab3 = c("Edat>91 anys", "Pulm?", "L"),
  colors = c("white", "grey"),
  forma = c("box", "box")
)
}
\arguments{
\item{pob_lab}{Pob lab}

\item{pob_lab1}{Pob lab1}

\item{pob_lab2}{Pob lab2}

\item{pob_lab3}{Pob lab3}

\item{pob}{Pob}

\item{pob1}{Pob1}

\item{exc1}{Exc1}

\item{exc_lab1}{Exc lab1}

\item{pob2}{Pob2}

\item{exc2}{Exc2}

\item{exc_lab2}{Exc lab2}

\item{pob3}{Pob3}

\item{exc3}{Exc3}

\item{exc_lab3}{Exc lab3}

\item{colors}{Colors}

\item{forma}{Forma}
}
\description{
Diagrama Flowchart 3 Grups
}
