% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exemple.R
\docType{data}
\name{exemple}
\alias{exemple}
\title{Base de dades iris i lloc}
\format{
Una data frame amb 150 files i 6 variables
\describe{
\item{Sepal.Length}{Sepal.Length ()}
\item{Sepal.Width}{Sepal.Width ()}
\item{Petal.Length}{Petal.Length ()}
\item{Petal.Width}{Petal.Width ()}
\item{Species}{Species ()}
\item{lloc}{lloc ()}
}
}
\usage{
exemple
}
\description{
Base de dades iris i lloc
}
\keyword{datasets}
