% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generar_flow_chart_consort.R
\name{Flow_chart_Consort}
\alias{Flow_chart_Consort}
\title{Dibuixa un Flow chart ggconsort}
\usage{
Flow_chart_Consort(
  dt,
  exclusions = c("declined", "prior_chemo", "bone_mets"),
  sequencial = TRUE,
  grups = NA,
  lab_start = "Assessed for eligibility",
  lab_random = "Analyzed sample",
  lab_exclusions = NULL
)
}
\arguments{
\item{dt}{Dataframe tibble}

\item{exclusions}{Vector de n strings amb la definicio del criteris:exclusions}

\item{sequencial}{Logic TRUE FALSE Exclusions sequencials o No}

\item{grups}{String referent al grup de dues categories (Opcional)}

\item{lab_start}{String referida a la etiqueta de l acaixa inici (default -Assessed for eligibility- )}

\item{lab_random}{String referida a la etiqueta de caixa post exclusions (default "Analyzed sample" )}

\item{lab_exclusions}{Vector etiquetes referent a les N exclusions (Opcional)}
}
\description{
Dibuixa un Flow chart ggconsort  a partir  llista de n criteris exclusio
}
\examples{

k<-Flow_chart_Consort(dt=ggconsort::trial_data,
                   exclusions=c("declined","prior_chemo","bone_mets"),
                   sequencial=TRUE,
                   grups=NA,
                   lab_start="Assessed for eligibility",
                   lab_random="Analyzed sample",
                   lab_exclusions=NULL)
k                    
}
