% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generar_flow_chart_consort.R
\name{Generar_flow_chart_consort}
\alias{Generar_flow_chart_consort}
\title{Dibuixa un flow_chart_ggconsort}
\usage{
Generar_flow_chart_consort(
  dt = iris,
  taulavariables = "conductor_cars.xls",
  camp = "camp",
  criteris = "exclusio",
  grups = NA,
  missings = F,
  sequencial = T,
  labels = NULL,
  lab_start = "Assessed for eligibility",
  lab_random = "Analyzed sample",
  ...
)
}
\arguments{
\item{dt}{Dataframe tibble}

\item{taulavariables}{String referent a path o tibble on hi ha una columna consten les exclusions: "conductor_cars.xls"}

\item{camp}{String referent al nom de la columna de les variables implicades (Inecesari)}

\item{criteris}{String referent a la columna on consten les exclusions}

\item{grups}{String referent al grup de dues categories}

\item{missings}{Logic TRUE FALSE (De moment no operatiu)}

\item{sequencial}{Logic TRUE FALSE.Exclusions sequencials o No}

\item{labels}{String nom del camp on hi ha les etiquetes de les exclusions (Optatiu)}

\item{lab_start}{String referida a la etiqueta de l acaixa d inici (default "Assessed for eligibility" )}

\item{lab_random}{String referida a la etiqueta de caixa post exclusions (default "Analyzed sample" )}

\item{...}{Altres parametres}
}
\description{
Dibuixa un flow_chart_ggconsort a partir d una llista de n criteris d exclusio que estan en un fitxer extern (Tipo excel)
}
\examples{
KK<-Generar_flow_chart_consort(
dt=iris,
taulavariables=conductor_cars,
camp="camp",
criteris="exclusio",
grups=NA,
missings=F,
sequencial=T,
labels="descripcio",
lab_start="Assessed for eligibility",
lab_random="Analyzed sample")

KK
}
