% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggflowchart2.R
\docType{package}
\name{ggflowchart2}
\alias{ggflowchart2}
\title{ggflowchart2 Package}
\description{
ggflowchart2 conté funcions per crear un diagrama de flux a partir de funcions de ggconsort i diagramma.
Si volem aplicar ggconsort,hem de actualitzar devtools::install_github("jrealgatius/ggflowchart").
}
\author{
Jordi Real,Ray Puig \email{-jordireal@gmail.com-}
}
