% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_flow_chart_diagrama.R
\name{criteris_exclusio}
\alias{criteris_exclusio}
\title{Aplica criteris partir d una llista de n criteris exclusio}
\usage{
criteris_exclusio(
  dt = dades,
  taulavariables = "VARIABLES_R3b.xls",
  criteris = "exclusio1",
  missings = T,
  ...
)
}
\arguments{
\item{dt}{dt}

\item{taulavariables}{taulavariables}

\item{criteris}{criteris}

\item{missings}{missings}

\item{...}{Altres parametres}
}
\description{
Aplica criteris partir d una llista de n criteris exclusio
}
