% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_flow_chart_diagrama.R
\name{criteris_exclusio_diagrama}
\alias{criteris_exclusio_diagrama}
\title{Criteris_exclusio_diagrama}
\usage{
criteris_exclusio_diagrama(
  dt = dades,
  taulavariables = "VARIABLES_R3b.xls",
  criteris = "exclusio1",
  pob_lab = c("Pob inicial", "Pob final"),
  etiquetes = "etiqueta_exclusio",
  ordre = "exc_ordre",
  grups = NA,
  sequencial = F,
  colors = c("white", "grey"),
  forma = c("ellipse", "box"),
  missings = T,
  ...
)
}
\arguments{
\item{dt}{Dt}

\item{taulavariables}{Taulavariables}

\item{criteris}{Criteris}

\item{pob_lab}{Pob labels}

\item{etiquetes}{Etiquetes}

\item{ordre}{Ordre}

\item{grups}{Grups}

\item{sequencial}{Sequencial}

\item{colors}{Colors}

\item{forma}{Forma}

\item{missings}{missings}

\item{...}{Altres parametres}
}
\description{
Criteris_exclusio_diagrama
}
\examples{

conductor_cars<-conductor_cars\%>\%select(-exclusio)
exclusio=c("<5","","","","setosa")
conductor_cars2<-data.frame(conductor_cars,exclusio)
exc_ordre=c("1","","","","2")
conductor_cars2<-data.frame(conductor_cars2,exclusio,exc_ordre)


k2<-criteris_exclusio_diagrama(
dt=iris,
taulavariables=conductor_cars2,
criteris="exc_pre",
ordre="exc_ordre",
grups=NA,
etiquetes="descripcio",
sequencial = TRUE,
pob_lab=c("Pob inicial","Pob final"),
colors=c("white","grey"),
forma=c("ellipse","box"))

k2

}
