% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteris_exclusio_taula.R
\name{criteris_exclusio_taula}
\alias{criteris_exclusio_taula}
\title{criteris_exclusio_taula}
\usage{
criteris_exclusio_taula(
  dt = dades,
  taulavariables = here::here("Conductor_CANA.xlsx"),
  criteris = "exclusio",
  ordre = NA,
  ...
)
}
\arguments{
\item{dt}{dt}

\item{taulavariables}{taulavariables}

\item{criteris}{criteris}

\item{ordre}{ordre}

\item{...}{Altres parametres}
}
\description{
criteris_exclusio_taula
}
