% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggflowchart2.R
\docType{package}
\name{ggflowchart2}
\alias{ggflowchart2}
\title{ggflowchart2 Package}
\description{
ggflowchart contains functions to create flow-chart from ggconsort functions.
}
\author{
Jordi Real,Ray Puig \email{-jordireal@gmail.com-}
}
